# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HotspotActivationRequest": {
        "additionalProperties": false,
        "description": "Body shared by `POST /api/hotspot/enable` and `/disable`.\n\nThese exist as their own routes so the UI's on/off switch never resends the\nwhole configuration \u2014 and therefore never has to be holding the passphrase\njust to flip a switch.",
        "properties": {
          "confirm_uplink_loss": {
            "default": false,
            "description": "Acknowledges that this call will drop a connection currently in use",
            "title": "Confirm Uplink Loss",
            "type": "boolean"
          }
        },
        "title": "HotspotActivationRequest",
        "type": "object"
      },
...
      "HotspotConfigRequest": {
        "additionalProperties": false,
        "description": "`PUT /api/hotspot` body \u2014 a full replace, not a merge.\n\nDeliberately not a PATCH. Merging a partial body would let a request that\nsays nothing about `hidden` silently inherit a previous `false`, which is\nexactly the security-relevant field an operator most expects to be\nexplicit. The passphrase is the single exception, and only because omitting\nit is the mechanism that keeps the secret write-only.",
        "properties": {
          "band": {
            "default": "bg",
            "description": "bg is 2.4 GHz, a is 5 GHz",
            "enum": [
              "bg",
              "a"
            ],
            "title": "Band",
            "type": "string"
          },
          "channel": {
            "default": 0,
            "description": "0 chooses automatically",
            "maximum": 196.0,
            "minimum": 0.0,
            "title": "Channel",
            "type": "integer"
          },
          "confirm_uplink_loss": {
            "default": false,
            "description": "Acknowledges that the chosen interface currently carries a connection which raising the hotspot will drop",
            "title": "Confirm Uplink Loss",
            "type": "boolean"
          },
          "enabled": {
            "default": false,
            "description": "Whether the hotspot should be running after this call",
            "title": "Enabled",
            "type": "boolean"
          },
          "gateway_cidr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The Pi's address on the hotspot network; omit to use the configured default",
            "title": "Gateway Cidr"
          },
          "hidden": {
            "default": true,
            "description": "Suppress SSID broadcast so clients must know the name in advance. Not a security control on its own",
            "title": "Hidden",
            "type": "boolean"
          },
          "interface": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9_.-]{1,15}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Wireless interface to use; omit to choose one automatically",
            "title": "Interface"
          },
          "passphrase": {
            "anyOf": [
              {
                "format": "password",
                "type": "string",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "description": "Omit to keep the currently stored password unchanged",
            "title": "Passphrase"
          },
          "security": {
            "default": "wpa2",
            "description": "wpa3 is experimental on Raspberry Pi radios",
            "enum": [
              "wpa2",
              "wpa3"
            ],
            "title": "Security",
            "type": "string"
          },
          "ssid": {
            "description": "The network name clients look for; 1-32 UTF-8 bytes",
            "title": "Ssid",
            "type": "string"
          }
        },
        "required": [
          "ssid"
        ],
        "title": "HotspotConfigRequest",
        "type": "object"
      },
      "HotspotErrorSummary": {
        "description": "The last control failure, kept so the UI can explain a hotspot that is not up.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ts": {
            "description": "Unix ms",
            "title": "Ts",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "message",
          "ts"
        ],
        "title": "HotspotErrorSummary",
        "type": "object"
      },
      "HotspotStateResponse": {
        "description": "`GET /api/hotspot` and every mutator's success body.\n\nNever 503s: a host that cannot do any of this answers 200 with\n`available: false`, matching how the app already degrades when `librtlsdr`\nis missing. A client can therefore always render *something*.",
        "properties": {
          "active": {
            "description": "Whether the hotspot is running right now",
            "title": "Active",
            "type": "boolean"
          },
          "auth_token_configured": {
            "description": "Whether SENTRY_AUTH_TOKEN is set",
            "title": "Auth Token Configured",
            "type": "boolean"
          },
          "available": {
            "description": "Whether access-point control works on this host",
            "title": "Available",
            "type": "boolean"
          },
          "band": {
            "default": "bg",
            "enum": [
              "bg",
              "a"
            ],
            "title": "Band",
            "type": "string"
          },
          "channel": {
            "default": 0,
            "minimum": 0.0,
            "title": "Channel",
            "type": "integer"
          },
          "configured": {
            "description": "Whether a hotspot profile exists",
            "title": "Configured",
            "type": "boolean"
          },
          "confirm_deadline_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Unix ms by which POST /api/hotspot/confirm must arrive",
            "title": "Confirm Deadline Ms"
          },
          "control_enabled": {
            "description": "Whether SENTRY_HOTSPOT_CONTROL_ENABLED is set",
            "title": "Control Enabled",
            "type": "boolean"
          },
          "enabled": {
            "description": "Whether the profile is set to come up on boot",
            "title": "Enabled",
            "type": "boolean"
          },
          "gateway_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The address a joined client should point Sentinel at, e.g. 10.42.0.1",
            "title": "Gateway Address"
          },
          "gateway_cidr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gateway Cidr"
          },
          "generated_at": {
            "description": "Unix ms",
            "title": "Generated At",
            "type": "integer"
          },
          "hidden": {
            "default": true,
            "title": "Hidden",
            "type": "boolean"
          },
          "interface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interface"
          },
          "last_error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HotspotErrorSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "passphrase_set": {
            "default": false,
            "description": "Whether a password is stored; never the password itself",
            "title": "Passphrase Set",
            "type": "boolean"
          },
          "pending_confirmation": {
            "default": false,
            "description": "A hotspot change is awaiting confirmation and will roll back without it",
            "title": "Pending Confirmation",
            "type": "boolean"
          },
          "security": {
            "default": "wpa2",
            "enum": [
              "wpa2",
              "wpa3"
            ],
            "title": "Security",
            "type": "string"
          },
          "ssid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssid"
          },
          "uplink_interface_is_hotspot_interface": {
            "default": false,
            "description": "True when raising the hotspot would drop this host's own connection",
            "title": "Uplink Interface Is Hotspot Interface",
            "type": "boolean"
          },
          "warnings": {
            "default": [],
            "items": {
              "enum": [
                "auth_token_missing",
                "advertised_host_overrides_gateway",
                "single_radio_uplink_loss",
                "nm_unavailable"
              ],
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "available",
          "control_enabled",
          "auth_token_configured",
          "configured",
          "enabled",
          "active",
          "generated_at"
        ],
        "title": "HotspotStateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Multi-dongle RTL-SDR controller for a Raspberry Pi.",
    "title": "Sentry",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/hotspot": {
      "delete": {
        "description": "Forget the network entirely, including its stored password.",
        "operationId": "delete_hotspot_api_hotspot_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Delete the hotspot configuration",
        "tags": [
          "hotspot"
        ]
      },
      "get": {
        "description": "Describe the hotspot, degrading rather than failing.\n\nAlways 200. A host with no NetworkManager answers `available: false` and a\n`nm_unavailable` warning \u2014 a client can therefore always render something\ntruthful instead of an error it cannot explain.",
        "operationId": "get_hotspot_api_hotspot_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotStateResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "The hotspot's current configuration and state",
        "tags": [
          "hotspot"
        ]
      },
      "put": {
        "description": "Write the whole configuration, then bring the hotspot up or down to match.\n\nOmitting `passphrase` keeps the stored one \u2014 the mechanism that lets an\noperator rename the network without the server ever handling the secret\nagain (WCAG 3.3.7 as much as security: not re-asking for something\nunchanged is a requirement, not a courtesy).",
        "operationId": "put_hotspot_api_hotspot_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotspotConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace the hotspot configuration",
        "tags": [
          "hotspot"
        ]
      }
    },
...
    "/api/hotspot/disable": {
      "post": {
        "description": "Bring the hotspot down and stop it starting on boot.",
        "operationId": "disable_hotspot_api_hotspot_disable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotspotActivationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop the hotspot",
        "tags": [
          "hotspot"
        ]
      }
    },
    "/api/hotspot/enable": {
      "post": {
        "description": "Bring the hotspot up provisionally; it rolls back unless confirmed.\n\nSeparate from `PUT` so the UI's on/off switch never resends the whole\nconfiguration, and therefore never has to be holding the passphrase just to\nflip a switch.",
        "operationId": "enable_hotspot_api_hotspot_enable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotspotActivationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start the configured hotspot",
        "tags": [
          "hotspot"
        ]
      }
    },
...
  }
}